requirements.txt
numpy
Pillow
python-dotenv
scipy
tensorflow
tensorflow-hub
wikipedia-api
